make syslinux package management optional
when finer control over the syslinux package is required, dedicated
puppet modules can be used. given that the tftp and the pxe module
would then end up managing the same package, one must not in order
for the catalog to compile.
this change make the syslinux package management optional, but keeps
the current behaviour for backwards compatibility.